fix: DB Type 맞게 변경 처리

+++ Projects/Data/Sources/DTO/DataBase/UserDTO.swift
@@ -13,23 +13,23 @@ import Domain
 public class UserDTO: NSManagedObject {
     @NSManaged public var startTime: Date
     @NSManaged public var finishTime: Date?
-    @NSManaged public var settingTime: Int
+    @NSManaged public var settingTime: Int32
     @NSManaged public var success: Bool
     @NSManaged public var concentrateMode: Bool
-    @NSManaged public var stopButtonClicked: Int
+    @NSManaged public var stopButtonClicked: Int16
     
     // Realm의 ObjectId 대신 UUID 사용
     @NSManaged public var id: UUID
     
     @discardableResult
     static func create(in context: NSManagedObjectContext,
-                      settingTime: Int) -> UserDTO {
+                      settingTime: Int32) -> UserDTO {
         let dto = UserDTO(context: context)
         dto.id = UUID()
         dto.startTime = Date()
         dto.success = false
         dto.concentrateMode = false
-        dto.settingTime = Int(settingTime)
+        dto.settingTime = settingTime
         dto.stopButtonClicked = 0
         return dto
     }
@@ -37,6 +37,6 @@ public class UserDTO: NSManagedObject {
 
 extension UserDTO {
     public var toDomain: UserEntity {
-        return UserEntity(id: id, startTime: startTime, finishTime: finishTime, settingTime: settingTime, success: success, concentrateMode: concentrateMode, stopButtonClicked: stopButtonClicked)
+        return UserEntity(id: id, startTime: startTime, finishTime: finishTime, settingTime: Int(settingTime), success: success, concentrateMode: concentrateMode, stopButtonClicked: Int(stopButtonClicked))
     }
 }
